+++ Lab2/src/include/Sorting.h
@@ -5,11 +5,12 @@
 #include "common/Macros.h"
 
 typedef int (*CompFunc)(const void *, const void *);
+
 typedef void (*SortFunc)(void *ptr, unsigned long count, unsigned long size, CompFunc);
 
 void choose_sort(void *ptr, unsigned long count, unsigned long size, CompFunc);
 void comb_sort(void *ptr, unsigned long count, unsigned long size, CompFunc);
 
-SortFunc getSort(int pos);
+SortFunc getSort(long pos);
 
 #endif //LAB2_SORTING_H
